backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
httpx==0.28.1
openai==1.58.1
pydantic==2.10.4
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
gspread==6.1.4
google-auth==2.38.0
stripe==11.6.0
resend==2.26.0
